+++ data/src/main/java/com/nadosunbae_android/data/mapper/mypage/MypageMapper.kt
@@ -4,8 +4,32 @@ import com.nadosunbae_android.data.model.response.mypage.ResponseMypageMyInfo
 import com.nadosunbae_android.domain.model.mypage.MyPageMyInfo
 import com.nadosunbae_android.domain.model.mypage.MyPageQuestionData
 import com.nadosunbae_android.data.model.response.mypage.ResponseMypageQuestionData
+import com.nadosunbae_android.domain.model.classroom.ClassRoomData
 
 object MypageMapper {
+
+    //마이페이지
+    fun mapperToMyPageQuestion(responseMypageQuestionData: MyPageQuestionData): List<ClassRoomData> {
+        return responseMypageQuestionData.data.classroomPostList.map {
+            ClassRoomData(
+                postId = it.postId,
+                title = it.title,
+                content = it.content,
+                createdAt = it.createdAt,
+                writer = ClassRoomData.Writer(
+                    nickname = it.writer.nickname,
+                    profileImageId = it.writer.profileImageId,
+                    writerId = it.writer.writerId
+                ),
+                isLiked = it.like.isLiked,
+                likeCount = it.like.likeCount,
+                commentCount = it.commentCount
+            )
+
+        }
+
+    }
+
     fun mapperToQuestion(responseMypageQuestionData: ResponseMypageQuestionData): MyPageQuestionData {
         return MyPageQuestionData(
             data = MyPageQuestionData.Data(
@@ -14,7 +38,10 @@ object MypageMapper {
                         commentCount = it.commentCount,
                         content = it.content,
                         createdAt = it.createdAt,
-                        likeCount = it.likeCount,
+                        MyPageQuestionData.Data.ClassroomPost.Like(
+                            isLiked = it.like.isLiked,
+                            likeCount = it.like.likeCount,
+                        ),
                         postId = it.postId,
                         title = it.title,
                         writer = MyPageQuestionData.Data.ClassroomPost.Writer(
