+++ data/src/main/java/com/nadosunbae_android/data/model/response/mypage/ResponseMypageQuestionData.kt
@@ -15,7 +15,7 @@ data class ResponseMypageQuestionData(
             val commentCount: Int,
             val content: String,
             val createdAt: Date?,
-            val likeCount: Int,
+            val like : Like,
             val postId: Int,
             val title: String,
             val writer: Writer
@@ -24,6 +24,11 @@ data class ResponseMypageQuestionData(
                 val nickname: String,
                 val profileImageId: Int,
                 val writerId: Int
+            )
+
+            data class Like(
+                val isLiked : Boolean,
+                val likeCount : Int
             )
         }
     }
