+++ domain/src/main/java/com/nadosunbae_android/domain/model/classroom/ClassRoomData.kt
@@ -9,6 +9,7 @@ data class ClassRoomData(
     val createdAt: Date?,
     val writer: Writer,
     val likeCount: Int,
+    val isLiked : Boolean,
     val commentCount: Int
 ) {
     data class Writer(
