+++ domain/src/main/java/com/nadosunbae_android/domain/model/mypage/MyPageQuestionData.kt
@@ -1,5 +1,6 @@
 package com.nadosunbae_android.domain.model.mypage
 
+import com.nadosunbae_android.domain.model.like.LikeData
 import java.util.*
 
 data class MyPageQuestionData(
@@ -13,7 +14,7 @@ data class MyPageQuestionData(
             val commentCount: Int,
             val content: String,
             val createdAt: Date?,
-            val likeCount: Int,
+            val like : Like,
             val postId: Int,
             val title: String,
             val writer: Writer
@@ -22,6 +23,11 @@ data class MyPageQuestionData(
                 val nickname: String,
                 val profileImageId: Int,
                 val writerId: Int
+            )
+
+            data class Like(
+                val isLiked : Boolean,
+                val likeCount : Int
             )
         }
     }
